fix(errors): address review comments — grammar, docs, oauth2 server_error, file rename
- Fix grammar in conflict / disk-full messages ("already exist" -> "already exists",
  "due the leak" -> "due to the lack")
- Update .agents docs to reference the new AuthupError classes instead of the
  removed apps/server-core/src/utils/errors.ts helpers
- Remove stray double-space in permission/service.ts import
- OAuth2Error.signingKeyMissing() now sets data.error = "server_error" instead
  of inheriting the "invalid_request" default; adds SERVER_ERROR to OAuth2ErrorCode
  per RFC 6749 §5.2
- Rename apps/server-core/src/core/oauth2/helpers/erorr.ts -> error.ts (pre-existing
  typo, only two import sites)
