+++ .agents/conventions.md
@@ -16,7 +16,7 @@
 ## Validation & Error Handling
 
 - **Validation**: `validup` framework with `@validup/adapter-zod` for Zod schema integration
-- **Errors**: `@authup/errors` provides `AuthupError` (extends `BaseError` from `@ebec/core`) plus domain subclasses with `Symbol.for(...)`-keyed duck-type guards (`isAuthupError`, `isOAuth2Error`, `isJWTError`, etc.). HTTP-status concern is decoupled — `ERROR_CODE_TO_STATUS` / `httpStatusFromCode(code)` map semantic codes to HTTP statuses in the adapter. **Core/domain code throws only `AuthupError` (or subclasses); `@ebec/http` classes are confined to the HTTP middleware for foreign-error translation.** New throw helpers in `apps/server-core/src/utils/errors.ts`: `entityNotFound`, `entityConflict`, `badRequest`, `forbidden`, `unauthorized`.
+- **Errors**: `@authup/errors` provides `AuthupError` (extends `BaseError` from `@ebec/core`) plus dedicated subclasses (`BadRequestError`, `UnauthorizedError`, `EntityNotFoundError`, `EntityConflictError`, `InternalError`, etc.) with `Symbol.for(...)`-keyed duck-type guards in sibling `check.ts` files (`isAuthupError`, `isOAuth2Error`, `isJWTError`, etc.). HTTP-status concern is decoupled — `ERROR_CODE_TO_STATUS` / `httpStatusFromCode(code)` map semantic codes to HTTP statuses in the adapter. **Core/domain code throws only `AuthupError` (or subclasses); `@ebec/http` classes are confined to the HTTP middleware for foreign-error translation.** Domain-specific identity workflow errors (e.g. `RegistrationDisabledError`, `PasswordRecoveryDisabledError`) live in their respective `core/identity/<workflow>/` folders.
 - **Validation location**: Validators from `@authup/core-kit` (e.g., `RoleValidator`, `UserValidator`) run inside core services, not in controllers. Services receive raw `Record<string, any>` data and call `validator.run(data, { group: ValidatorGroup.CREATE })` internally. Controllers use `useRequestBody(req)` to pass the raw body to the service.
 - **Canonical identifier form**: `name` (every entity) and `user.email` are stored as `LOWER(TRIM(value))`. New `name`-style columns must chain `.trim().toLowerCase()` in their validator before the format check, and use `=` (not `LIKE`) in repository lookups. See `.agents/architecture.md#canonical-identifier-form` for the full rationale.
