+++ apps/server-core/src/core/oauth2/helpers/error.ts
