+++ apps/server-core/src/core/oauth2/helpers/index.ts
@@ -5,4 +5,4 @@
  *  view the LICENSE file that was distributed with this source code.
  */
 
-export * from './erorr.ts';
+export * from './error.ts';
